third_party/WebKit/LayoutTests/shadow-dom/shadowroot-mode.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="host1"></div>
<div id="host2"></div>
<div id="host3"></div>
<input id="input">
<script>

test(function() {
  host1.attachShadow({mode : 'open'});
  var closedShadowRoot = host2.attachShadow({mode : 'closed'});
  host3.createShadowRoot();
  assert_equals(host1.shadowRoot.mode, 'open');
  assert_equals(closedShadowRoot.mode, 'closed');
  assert_equals(host3.shadowRoot.mode, 'open');
  // Note: mode for V0 shadow and UA shadow is not specified anywhere,
  // Blink returns reasonable default values for these.
  assert_equals(internals.shadowRoot(input).mode, 'closed');
}, 'ShadowRoot.mode should return open or closed, and only open if v0.');
</script>
